pyproject.toml
[project]
name = "my-todo-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "google-generativeai>=0.8.5",
    "openai-agents>=0.0.19",
]
